Gleb/add/tracking sheet endpoint

+++ app/models/crew/project_crew.py
@@ -6,8 +6,8 @@ class ProjectCrew(TimestampMixin, ModelMixin, db.Model):
 
     __tablename__ = "project_crew"
 
-    project_crew_id = db.Column(db.Integer, primary_key=True)
-    project_id = db.Column(db.Integer, nullable=False)
+    project_crew_id = db.Column(db.BigInteger, primary_key=True)
+    project_id = db.Column(db.BigInteger, nullable=False)
     crew_id = db.Column(db.Integer, nullable=False)
 
     crew = db.relationship(
